+++ mapsforge-map/src/main/java/org/mapsforge/map/model/Model.java
@@ -16,22 +16,9 @@
  */
 package org.mapsforge.map.model;
 
-import org.mapsforge.map.model.common.Persistable;
-import org.mapsforge.map.model.common.PreferencesFacade;
-
-public class Model implements Persistable {
+public class Model {
     public final DisplayModel displayModel = new DisplayModel();
     public final FrameBufferModel frameBufferModel = new FrameBufferModel();
     public final MapViewDimension mapViewDimension = new MapViewDimension();
     public final MapViewPosition mapViewPosition = new MapViewPosition(displayModel);
-
-    @Override
-    public void init(PreferencesFacade preferencesFacade) {
-        this.mapViewPosition.init(preferencesFacade);
-    }
-
-    @Override
-    public void save(PreferencesFacade preferencesFacade) {
-        this.mapViewPosition.save(preferencesFacade);
-    }
 }