+++ docs/Changelog.md
@@ -14,6 +14,7 @@
 - Remove `IMapViewPosition`, use `MapViewPosition` [#1591](https://github.com/mapsforge/mapsforge/pull/1591)
 - Remove `Filter`, use `ThemeCallback` [#1573](https://github.com/mapsforge/mapsforge/pull/1573)
 - Move `MapViewerTemplate` in samples [#1616](https://github.com/mapsforge/mapsforge/pull/1616)
+- Move preferences in samples [#1617](https://github.com/mapsforge/mapsforge/pull/1617)
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aclosed+milestone%3A0.23.0)